Make post dates optional, add XHTML format support
This adds a format command that allows the use of either CommonMark
or raw XHTML in posts. It also makes the date command optional; if
a date is not specified, then only a permalink page is generated and
the post does not appear in any other generated list.

The test suite has also been re-enabled (it was accidentally disabled
in the Maven build).

Fix #4
Affects #3