Add format providers
This modularizes the codebase further such that blog post formats
are now provided via OSGi or ServiceLoader services. It moves the
CommonMark format into a separate optional module and adds a new
configuration option to specify the default format of posts.  It also
adds a command-line option to list the available formats.

Fix #5